+++ src/client.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
+import Header from './components/header/header';
+import Footer from './components/footer/footer';
 import App from './components/app/app';
 import StyleGuide from './components/style-guide/style-guide';
+import Changelog from './components/changelog/changelog';
 
 const applicationConfig = window.applicationConfig;
 
@@ -12,6 +15,14 @@ const component = () => {
   switch (renderPage) {
     case '/style-guide':
       return <StyleGuide />;
+    case '/changelog':
+      return (
+        <div>
+          <Header />
+          <Changelog content={applicationConfig.changelogMarkdown} />
+          <Footer appVersion={applicationConfig.appVersion} />
+        </div>
+      );
     default:
       return <App version={applicationConfig.appVersion} />;
   }