+++ src/components/header/header.jsx
@@ -4,8 +4,10 @@ import styles from './header.less';
 
 export default function Status() {
   return (
-    <div className={styles['container']}>
-      <div className={styles['title']}>DJ Lama</div>
+    <div className={styles['spacing']}>
+      <div className={styles['container']}>
+        <div className={styles['title']}>DJ Lama</div>
+      </div>
     </div>
   );
 }